refactor: rename components to use PascalCase for consistency
